Fix assert when reliable channel reactivates
When a reliable channel is dormant (no messages sent) for an extended
period of time, it will trigger an assert when it reactivates (the next
time you send a message). This occurs because reliable's sequence number
will continue to increment, and the SequenceBuffer will assume the
packet is too old when we try to insert it. Since we're sending, we know
that's not true, so this adds a new `InsertAndIgnoreAge` method to
`SequenceBuffer` and converts the reliable channel sending to use that
instead.

Fixes #138 and #146.