Remove `SentPacketEntry::acked`
`SentPacketEntry::acked` never had a `1` written to it, so the code that
checked for it was no-op. This removes the field and instead removes the
acked packet from the `m_sentPackets`, so that we do not double-process
a packet (which we would do previously, but it would just be a no-op
since the packets were already removed from the `m_messageSendQueue`.