Fix for YOJIMBO_NEW() null pointer dereference.
YOJIMBO_DECLARE_MESSAGE_TYPE() expands to:
```message = YOJIMBO_NEW( allocator, msg_action );
    if ( !message )
        return NULL;```

But YOJIMBO_NEW() never checks the result from Allocate(), it just
immediately passes it to placement new(), which dereferences a null
pointer if Allocate() fails. This makes it impossible to gracefully
handle OOM situations, and leads to unpreventable crashes.

This change replaces the existing macro with a short lambda evaluated
on the spot that checks for the null pointer and avoids placement new
in OOM situations.